Replace @ in events to avoid notifying @mentions
When using Ansible to create Datadog monitors, the events we create here
could contain @mentions meant to appear in the monitor message. That can
cause teams meant to be notified when the monitor goes off to be notified
when the monitor is created, because of the event containging an @mention.

This replaces @mentions with (@)mentions, preventing the notification.